chore: add separate class for ResetEntries class implementation
